requirements.txt
absl-py==1.4.0
astor==0.8.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2023.7.22
charset-normalizer==2.0.12
dataclasses==0.8
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==2.10.0
idna==3.4
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ISR==2.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.7
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.6
pyaml==23.5.8
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyYAML==6.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
tqdm==4.64.1
typing_extensions==4.1.1
urllib3==1.26.16
Werkzeug==2.0.3
wrapt==1.15.0
zipp==3.6.0
